SEQNG-768 Some refactorings
