+++ modules/seqexec/web/client/src/test/scala/seqexec/web/client/model/ArbitrariesWebClient.scala
@@ -147,9 +147,11 @@ trait ArbitrariesWebClient extends ArbObservation with TableArbitraries {
     }
 
   implicit val cqtCogen: Cogen[CalibrationQueueTab] =
-    Cogen[(TableState[StepsTable.TableColumn], BatchExecState, Option[Observer])].contramap { x =>
-      (x.tableState, x.state, x.observer)
-    }
+    Cogen[
+      (TableState[StepsTable.TableColumn], BatchExecState, Option[Observer])]
+      .contramap { x =>
+        (x.tableState, x.state, x.observer)
+      }
 
   implicit val arbQueueSeqOperations: Arbitrary[QueueSeqOperations] =
     Arbitrary {
@@ -160,7 +162,6 @@ trait ArbitrariesWebClient extends ArbObservation with TableArbitraries {
 
   implicit val sopCogen: Cogen[QueueSeqOperations] =
     Cogen[RemoveSeqQueue].contramap(x => x.removeSeqQueue)
-
 
   implicit val arbInstrumentSequenceTab: Arbitrary[InstrumentSequenceTab] =
     Arbitrary {
@@ -362,9 +363,10 @@ trait ArbitrariesWebClient extends ArbObservation with TableArbitraries {
         g <- arbitrary[Boolean]
         i <- arbitrary[Option[Instrument]]
         d <- arbitrary[Option[Observation.Id]]
+        c <- arbitrary[Boolean]
         s <- arbitrary[TabSelected]
         l <- arbitrary[SectionVisibilityState]
-      } yield SequenceTabContentFocus(g, i, d, s, l)
+      } yield SequenceTabContentFocus(g, i, d, c, s, l)
     }
 
   implicit val stcfCogen: Cogen[SequenceTabContentFocus] =
